Add FindGpgme.cmake for cmake build
To support the cmake build, we need to find our dependencies. Generally
on cmake this is done with the use of find_package(foo). Then cmake will
try to locate a file FindFoo.cmake and run it in order to actually find
foo (basically the cmake equivalent of pkg-config). So for find_modules
that don't yet exist in the standard cmake installation we need to add
our own. This file, FindGpgme.cmake was copied from here:
https://github.com/rpm-software-management/librepo/blob/master/cmake/Modules/FindGpgme.cmake